requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.4
djangorestframework==3.16.0
dotenv==0.9.9
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2

drf-spectacular~=0.28.0
pandas~=2.3.1asgiref==3.9.1
attrs==25.3.0
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
